+++ libs/react/src/components/pages/survey/survey.tsx
@@ -5,6 +5,7 @@ import { cloneDeep, isEmpty } from 'lodash';
 import useSWR from 'swr';
 import { axiosFetcher } from '@coldpbc/fetchers';
 import { GlobalSizes } from '@coldpbc/enums';
+import { useSearchParams } from 'react-router-dom';
 
 export interface SurveyProps {
   surveyName: string;
@@ -24,8 +25,13 @@ export const Survey = (props: SurveyProps) => {
     [`/surveys/${surveyName}`, 'GET'],
     axiosFetcher,
   );
-
+  const [searchParams, setSearchParams] = useSearchParams();
   const submitSurvey = () => {
+    const param = searchParams.get('surveyName');
+    if (param) {
+      searchParams.delete('surveyName');
+      setSearchParams(searchParams);
+    }
     setShow(false);
   };
 
@@ -48,6 +54,15 @@ export const Survey = (props: SurveyProps) => {
         });
       }
     }
+  };
+
+  const onSurveyClose = () => {
+    const param = searchParams.get('surveyName');
+    if (param) {
+      searchParams.delete('surveyName');
+      setSearchParams(searchParams);
+    }
+    setShow(false);
   };
 
   useEffect(() => {
@@ -104,6 +119,7 @@ export const Survey = (props: SurveyProps) => {
           dismiss: {
             label: 'close',
             dismissible: true,
+            onClick: onSurveyClose,
           },
         }}
       >
@@ -125,6 +141,7 @@ export const Survey = (props: SurveyProps) => {
             dismiss: {
               label: 'close',
               dismissible: true,
+              onClick: onSurveyClose,
             },
           }}
         >